# src/main/java/se/sundsvall/myrepresentative/api/model/GrantorDetails.java
package se.sundsvall.myrepresentative.api.model;

import io.swagger.v3.oas.annotations.media.Schema;
import se.sundsvall.dept44.common.validators.annotation.ValidUuid;
import se.sundsvall.myrepresentative.config.Builder;

@Builder
@Schema(description = "Mandate grantor model")
public record GrantorDetails(

	@Schema(description = "The name of the granting organization or person", example = "Ankeborgs Margarinfabrik") String name,

	@ValidUuid @Schema(description = "The partyId of the issuing organization or person", example = "fb2f0290-3820-11ed-a261-0242ac120002") String grantorPartyId,

	@ValidUuid @Schema(description = "PartyId of the issuing person / signatory", example = "fb2f0290-3820-11ed-a261-0242ac120003") String signatoryPartyId) {
}